Openclsim integration into HALEM
